+++ test/rules/no-lonely-if.spec.js
@@ -1,0 +1,15 @@
+'use strict';
+
+// eslint no-lonely-if: "error"
+
+let condition = false; let anotherCondition = true;
+
+if (condition) {
+  // ...
+}
+else if (anotherCondition) {
+  // ...
+}
+else {
+  // ...
+}