+++ test/rules/no-unused-vars.spec.js
@@ -1,5 +1,7 @@
 'use strict';
 
+// eslint no-unused-vars
+
 (function (foo, bar, baz) {
   return bar + foo + baz;
 })();